Separate workgroup and org registry to separate modules
